==> code/acquire_data/03a_event_data_exploration.Rmd <==
---
title: "event_data_exploration"
author: "Hennes"
date: "2024-12-18"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Libraries and Data
```{r libraries-and-data}
library(tidyverse)
library(sf)
library(ggmap)
library(ggspatial)
source("helpers.R")

df <- read_csv("../data/ACLED_Ukraine_2013-11-01-2024-12-16.csv")
```

# Variables
```{r variables}
names(df)
```

# Actors
```{r top-actors-1}
show_top(df, actor1, 30)
```

```{r top-actors-2}
show_top(df, actor2, 30)
```
# Geo Precision

```{r geo-precision}
show_top(df, geo_precision)
```

# Event Type
```{r event-type}
show_top(df, event_type)
``` 

# Sub Event Type
```{r sub-event-type}
show_top(df, sub_event_type)
``` 

# Plot the Frequency of Number of Fatalities
```{r plot-fatalities}
df %>% ggplot(aes(fatalities)) +
  scale_y_log10() +
  geom_histogram(binwidth = 1) +
  labs(title = "Frequency of number of fatalities", x = "Number of fatalities", y = "Frequency") +
  coord_cartesian(xlim = c(0, 200))
```
